Separate out device information functions
